websockets noticias version vieja - websockets actividades version nueva

+++ app/Http/Controllers/InvestigadoresController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Events\InvestigadorEvent;
 use App\Investigador;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 
 class InvestigadoresController extends Controller
@@ -15,51 +17,57 @@ class InvestigadoresController extends Controller
 
     public function store(Request $request)
     {
+        $this->validate(
+            $request,
+            ['email' => 'unique:investigadors,email'],
+            ['email.unique' => 'Ya existe un investigador con ese Email']
+        );
         $imagen = $this->agregarImagen($request->imagen);
-
-        $investigador = new Investigador();
-        $investigador->nombres = ucfirst($request->nombres);
-        $investigador->apellidos = ucfirst($request->nombres);
-        $investigador->publicar = $request->publicar;
-        $investigador->email = $request->email;
-        $investigador->nivel_estudio = ucfirst($request->nivel_estudio);
-        $investigador->cargo = ucfirst($request->cargo);
-        $investigador->imagen = $imagen['ruta'];
-        $investigador->imagenPublica = $imagen['rutaPublica'];
-        $investigador->save();
-
+        $datos = [
+            'nombres' => ucfirst($request->nombres), 'apellidos' => ucfirst($request->apellidos),
+            'publicar' => $request->publicar, 'email' => $request->email,
+            'nivel_estudio' => ucfirst($request->nivel_estudio), 'cargo' => ucfirst($request->cargo),
+            'imagen' => $imagen['ruta'], 'imagenPublica' => $imagen['rutaPublica']
+        ];
+        $investigador =  Investigador::create($datos);
+        broadcast(new InvestigadorEvent($investigador, 'agregar'))->toOthers();
         return $investigador;
     }
 
     public function update(Request $request, $id)
     {
         $investigador = Investigador::find($id);
-
-        if ($request->imagen != $investigador->imagen) {
-            //eliminar imagen vieja
-            Storage::disk('local')->delete($investigador->imagen);
-            //agregar imagen nueva
-            $imagen = $this->agregarImagen($request->imagen);
-
-            $investigador->imagen = $imagen['ruta'];
-            $investigador->imagenPublica = $imagen['rutaPublica'];
+        $investigador1 = Investigador::where('email', $request->email)->first();
+        if (is_null($investigador1) || $investigador->id == $investigador1->id) {
+            $datos = [];
+            if ($request->imagen != $investigador->imagen) {
+                //eliminar imagen vieja
+                Storage::disk('local')->delete($investigador->imagen);
+                //agregar imagen nueva
+                $imagen = $this->agregarImagen($request->imagen);
+                $datos += ['imagen' => $imagen['ruta'], 'imagenPublica' => $imagen['rutaPublica']];
+            }
+            $datos += [
+                'nombres' => ucfirst($request->nombres), 'apellidos' => ucfirst($request->apellidos),
+                'publicar' => $request->publicar, 'email' => $request->email,
+                'nivel_estudio' => ucfirst($request->nivel_estudio), 'cargo' => ucfirst($request->cargo)
+            ];
+            $investigador->update($datos);
+            broadcast(new InvestigadorEvent($investigador, 'editar'))->toOthers();
+            return $investigador;
+        } else {
+            return response()->json([
+                'errors' =>
+                ['email' => ['Ya existe un investigador con ese Email']]
+            ], 422);
         }
-
-        $investigador->nombres = ucfirst($request->nombres);
-        $investigador->apellidos = ucfirst($request->nombres);
-        $investigador->email = $request->email;
-        $investigador->nivel_estudio = ucfirst($request->nivel_estudio);
-        $investigador->publicar = $request->publicar;
-        $investigador->cargo = ucfirst($request->cargo);
-        $investigador->save();
-
-        return $investigador;
     }
 
     public function destroy($id)
     {
         $investigador = Investigador::find($id);
         Storage::disk('local')->delete($investigador->imagen);
+        broadcast(new InvestigadorEvent($investigador, 'eliminar'))->toOthers();
         $investigador->delete();
         return $investigador;
     }
@@ -68,7 +76,7 @@ class InvestigadoresController extends Controller
     {
         $imagen_array = explode(",", $imagen);
         $data = base64_decode($imagen_array[1]);
-        $image_name = time() . '.png';
+        $image_name = Auth::user()->id . '-' . rand(Auth::user()->id, 1000) . '-' . time() . '.png';
         Storage::disk('local')->put('/public/investigadores/' . $image_name, $data);
         $ruta = '/public/investigadores/' . $image_name;
         $rutaPublica = '/storage/investigadores/' . $image_name;
@@ -78,8 +86,8 @@ class InvestigadoresController extends Controller
     public function publicar(Request $request, $id)
     {
         $investigador = Investigador::find($id);
-        $investigador->publicar = $request->publicar;
-        $investigador->save();
+        $investigador->update(['publicar' => $request->publicar]);
+        broadcast(new InvestigadorEvent($investigador, 'editar'))->toOthers();
         return $investigador;
     }
 }
